+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-### 20140930
-* Initial release
-* Added Metro for Steam Skin
+### 20140930-3
+* Added PixelVision Skin
+
+### 20140930-2
 * Added Spanish language
 * Added license agreement
 * Somes fixes when checking Steam installation path
+
+### 20140930-1
+* Added Metro for Steam Skin
+
+### 20140930-0
+* Initial release